pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lexiguard-env"
version = "1.0.0"
description = "LexiGuard Legal Contract Review OpenEnv Environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "lexiguard.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["lexiguard", "lexiguard.server"]
package-dir = { "lexiguard" = ".", "lexiguard.server" = "server" }